moved everything but main into a subfolder called modules
